Take as exit candidates the maximum potential exits
Instead of start taking as exit candidates de relays with exit policy
and later filter them those relays that have also EXIT flag, take
either the set of relays with EXIT flag or exit policy, whatever is
greater and later restrict that set depending on the module and/or
parameters.
Also add debug logging lines to show which relays are being selected
and when.

Closes #35.